fix test to use only a single byte
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_4x@1496637 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/test/org/apache/lucene/index/TestMaxTermFrequency.java
@@ -107,7 +107,7 @@ public class TestMaxTermFrequency extends LuceneTestCase {
 
     @Override
     public long encodeNormValue(float f) {
-      return (long) f;
+      return (byte) f;
     }
 
     @Override
